Move provisioner into subdirectory
